Rename Mezzio Adapter to Mezzio Template Adapter
