+++ examples/usages/mezzio/config/config.php
@@ -16,7 +16,7 @@ $cacheConfig = [
 $aggregator = new ConfigAggregator([
     \Schranz\Templating\Integration\Mezzio\Blade\ConfigProvider::class,
     \Schranz\Templating\Integration\Mezzio\Latte\ConfigProvider::class,
-    \Schranz\Templating\Integration\Mezzio\Mezzio\ConfigProvider::class,
+    \Schranz\Templating\Integration\Mezzio\MezzioTemplate\ConfigProvider::class,
     \Schranz\Templating\Integration\Mezzio\Smarty\ConfigProvider::class,
     \Schranz\Templating\Integration\Mezzio\Mustache\ConfigProvider::class,
     \Schranz\Templating\Integration\Mezzio\Handlebars\ConfigProvider::class,
