+++ examples/usages/mezzio/config/routes.php
@@ -44,7 +44,7 @@ return static function (Application $app, MiddlewareFactory $factory, ContainerI
     $app->get('/handlebars', App\Handler\TemplateHandlebarsHandler::class, 'handlebars');
     $app->get('/mustache', App\Handler\TemplateMustacheHandler::class, 'mustache');
     $app->get('/smarty', App\Handler\TemplateSmartyHandler::class, 'smarty');
-    $app->get('/mezzio', App\Handler\TemplateMezzioHandler::class, 'mezzio');
+    $app->get('/mezzio-template', App\Handler\TemplateMezzioTemplateHandler::class, 'mezzio');
     $app->get('/latte', App\Handler\TemplateLatteHandler::class, 'latte');
     $app->get('/blade', App\Handler\TemplateBladeHandler::class, 'blade');
     $app->get('/api/ping', App\Handler\PingHandler::class, 'api.ping');
