+++ examples/usages/mezzio/src/App/src/ConfigProvider.php
@@ -43,7 +43,7 @@ class ConfigProvider
                 Handler\TemplateHandlebarsHandler::class => Handler\TemplateHandlebarsHandler::class,
                 Handler\TemplateMustacheHandler::class => Handler\TemplateMustacheHandler::class,
                 Handler\TemplateSmartyHandler::class => Handler\TemplateSmartyHandler::class,
-                Handler\TemplateMezzioHandler::class => Handler\TemplateMezzioHandler::class,
+                Handler\TemplateMezzioTemplateHandler::class => Handler\TemplateMezzioTemplateHandler::class,
                 Handler\TemplateLatteHandler::class => Handler\TemplateLatteHandler::class,
                 Handler\TemplateBladeHandler::class => Handler\TemplateBladeHandler::class,
                 Handler\PingHandler::class => Handler\PingHandler::class,
@@ -55,7 +55,7 @@ class ConfigProvider
                 Handler\TemplateHandlebarsHandler::class => Handler\TemplateHandlebarsHandlerFactory::class,
                 Handler\TemplateMustacheHandler::class => Handler\TemplateMustacheHandlerFactory::class,
                 Handler\TemplateSmartyHandler::class => Handler\TemplateSmartyHandlerFactory::class,
-                Handler\TemplateMezzioHandler::class => Handler\TemplateMezzioHandlerFactory::class,
+                Handler\TemplateMezzioTemplateHandler::class => Handler\TemplateMezzioTemplateHandlerFactory::class,
                 Handler\TemplateLatteHandler::class => Handler\TemplateLatteHandlerFactory::class,
                 Handler\TemplateBladeHandler::class => Handler\TemplateBladeHandlerFactory::class,
             ],
